fiz uma renca de coisa, adicionei o handler do Provider, adicionei uma parte das validaçoes e etc

+++ Wms.Core.Application/Commands/EntityCommands/DistributionCenterCommand/Delete/DistributionCenterDeleteHandler.cs
@@ -1,7 +1,5 @@
 using ErrorOr;
-using MapsterMapper;
 using MediatR;
-using Wms.Core.Domain.Entities.Entity;
 using Wms.Core.Infrastructure.Interfaces.EntityRepositoryInterface;
 
 namespace Wms.Core.Application.Commands.EntityCommands.DistributionCenterCommand.Delete;
@@ -9,18 +7,17 @@ namespace Wms.Core.Application.Commands.EntityCommands.DistributionCenterCommand
 public class DistributionCenterDeleteHandler : IRequestHandler<DistributionCenterDeleteCommand, Error?>
 {
     readonly IDistributionCenterRepository _repository;
-    readonly IMapper _mapper;
 
-
-    public DistributionCenterDeleteHandler(IDistributionCenterRepository repository, IMediator mediator, IMapper mapper)
+    public DistributionCenterDeleteHandler(IDistributionCenterRepository repository)
     {
         _repository = repository;
-        _mapper = mapper;
     }
 
     public async Task<Error?> Handle(DistributionCenterDeleteCommand command, CancellationToken cancellationToken)
     {
-        var cd = _mapper.Map<DistributionCenter>(command);
+        var cd = await _repository.GetByCode(command.Code);
+
+        if (cd is null) return null;
 
         await _repository.Delete(cd);
